Tidy up ActionTranspiler functional test
