+++ tests/Functional/Identifier/IdentifierTranspilerTest.php
@@ -17,7 +17,6 @@ use webignition\BasilTranspiler\Model\UseStatement;
 use webignition\BasilTranspiler\Model\UseStatementCollection;
 use webignition\BasilTranspiler\Model\VariablePlaceholderCollection;
 use webignition\BasilTranspiler\Tests\Functional\AbstractTestCase;
-use webignition\BasilTranspiler\Tests\Services\ExecutableCallFactory;
 use webignition\BasilTranspiler\VariableNames;
 use webignition\SymfonyDomCrawlerNavigator\Model\ElementLocator;
 use webignition\SymfonyDomCrawlerNavigator\Model\LocatorType;
@@ -26,27 +25,16 @@ use webignition\WebDriverElementCollection\WebDriverElementCollectionInterface;
 
 class IdentifierTranspilerTest extends AbstractTestCase
 {
-    const DOM_CRAWLER_NAVIGATOR_VARIABLE_NAME = '$domCrawlerNavigator';
-    const VARIABLE_IDENTIFIERS = [
-        VariableNames::DOM_CRAWLER_NAVIGATOR => self::DOM_CRAWLER_NAVIGATOR_VARIABLE_NAME,
-    ];
-
     /**
      * @var IdentifierTranspiler
      */
     private $transpiler;
-
-    /**
-     * @var ExecutableCallFactory
-     */
-    private $executableCallFactory;
 
     protected function setUp(): void
     {
         parent::setUp();
 
         $this->transpiler = IdentifierTranspiler::createTranspiler();
-        $this->executableCallFactory = ExecutableCallFactory::createFactory();
     }
 
     /**
